Remove algorithms module and move fallback to finalizers.
