--- pipeline/matching/algorithms/fallback/README.md
@@ -1,3 +1,0 @@
-# Fallback Matching Algorithm
-
-The fallback algorithm is designed to be a final step in the matching engine, to ensure that every user gets a match, preferably one they have not been matched to recently.
